Создание блока с погодой ( сохранение второе)
