+++ src/index.html
@@ -13,11 +13,11 @@
     <img>
     <div class="info-weather-box">
       <span class="degrees-weather">4</span>
-      <div class="detailed-weather">
+      <div class="detailed-weather text-uppercase">
         <span class="description-weather">Дождь например</span>
         <span class="wind-speed"></span>
-        <span class="precipitation"></span>
         <span class="humidity"></span>
+        <span class="precipitation "></span>
       </div>
     </div>
   </div>
