+++ src/styles/style.css
@@ -1,17 +1,35 @@
+@font-face {
+  font-family: "Roboto Light";
+  src: url('https://fonts.googleapis.com/css2?family=Roboto:wght@100&display=swap');
+}
+
 body {
+    font-size: 24px;
+    font-weight: bold;
     display: flex;
     flex-direction: row;
     align-items: center;
     justify-content: center;
+    font-family: "Roboto Light", sans-serif;
 }
 
 .info-weather-box{
   display: flex;
 }
 
+/*.location-weather {*/
+/*  font-size: 34px;*/
+/*  font-weight: bold;*/
+/*}*/
+
 .detailed-weather{
   display: flex;
   flex-direction: column;
+  font-size: 34px;
+}
+
+.degrees-weather{
+  font-size: 280px;
 }
 
 .map-web {
